+++ lib/mystic/extensions.rb
@@ -1,5 +1,7 @@
 #!/usr/bin/env ruby
 
+require "pathname"
+
 module Kernel
 	def self.silent
 		v = $VERBOSE
@@ -11,35 +13,20 @@ module Kernel
 end
 
 class String
-  def capitalize_first!
-    self[0] = self[0].upcase
-  end
-  
-  def desnake
-    downcase.split("_").map(&:capitalize)*' '
-  end
-	
   def escape
     Mystic.escape(self).untaint
   end
 	
-  alias_method :sanitize, :escape
-  
-	def truncate(len)
-		self[0..len-1]
-	end
-	
-	def sql_terminate
-		return self + ";" unless dup.strip.end_with? ";"
-		self
+	def terminate term=";"
+    strip.end_with?(term) ? self + term : self
 	end
 end
 
 class Array
-  def merge_keys(*keys)
+  def merge_keys *keys
 		raise ArgumentError, "No keys to merge." if keys.nil? || keys.empty?
-    raise ArgumentError, "Argument array must have the same number of elements as self." if keys.count != self.count
-    Hash[each_with_index.map{ |v,i| [keys[i],v] }]
+    raise ArgumentError, "Argument array must have the same number of elements as self." if keys.count != count
+    Hash[each_with_index.map { |v,i| [keys[i],v] }]
   end
 	
 	def symbolize
@@ -54,7 +41,7 @@ class Array
 		map { |o|
 			case o
 			when String
-				"'#{o.sanitize}'"
+				"'#{o.escape}'"
 			when Numeric
 				o.to_s
 			end
@@ -67,18 +54,6 @@ class Hash
 		Hash[values_at(*keys).merge_keys(*keys).reject{ |k,v| v.nil? }]
 	end
 	
-  def parify delim=" "
-    map { |pair| pair * delim }
-  end
-	
-	def compact
-		reject { |k,v| v.nil? }
-	end
-	
-	def compact!
-		reject! { |k,v| v.nil? }
-	end
-	
 	def symbolize
 		Hash[map { |k,v| [k.to_sym, v]}]
 	end
@@ -88,7 +63,7 @@ class Hash
 	end
   
   def sqlize
-    reject { |k,v| v.nil? || (v.empty? rescue false) }.map{ |k,v| "#{k}=#{Integer === v ? v : "'#{v.to_s.sanitize}'" }" }
+    reject { |k,v| v.nil? || (v.empty? rescue false) }.map { |pair| pair.sqlize.join '=' }
   end
 end